services/plugin-state-manager/requirements.txt
# services/plugin-state-manager/requirements.txt
# FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# HTTP client
httpx>=0.27.0

# Redis (caching)
redis==5.0.1

# Authentication & Security
pyjwt==2.8.0
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Utilities
uuid
pyyaml==6.0.1
